Average over all given realizations
We added extra maps that allow us to parse data over the entire space of
possible {chemical potentials} x {realizations} for a given temperature,
and then reduce (average) over the realization subspaces. This greatly
reduces the error if multiple realizations are available per data set,
as the errors are summed in quadrature and divided by N.

The implementation using a struct full of std::maps is very
uncomfortable, as it feels like we waste a lot of time looking up values
in each individual map. In the future we should think about rewriting
the code to reverse the ordering of the struct and outermost maps.