Move chemical potential map outside the structs
This refactors the data structs to now be wrapped inside a std::map
sending chemical potentials to an image struct of variables, rather than
each struct containing a map for each individual variable. This reduces
the number of lookups we have to do to perform calculations for the
averages and parsing, but should still preserve (or even improve)
memory/cache performance, since the vectors we loop over are still
contiguous within a given struct.