Add const correctness to map data structures
Before we start adding more and more layers to the program, we *need* to
address const correctness: once we create the unaveraged map/struct, the
members and map are no longer meant to be modified by the parsing and
averaging functions. All computations are done with intermediates using
the <numeric> library and stored directly in a newly constructed map, so
the original map should not change.

This prevents us from accidentally modifying source data entries during
the averaging and printing processes, though it *does not* prevent us
from making computational or logic errors.