Add qmc::quantity backend, refactor data handling
This adds support for simultaneous parsing of both mu>0 and mu<0 data,
corresponding to n>1 and n<1. For n>1 (above half-filling), the
correlation function data is calculated using the inverted operators
n' = 2-n, h' = n, and d' = 1-n+d So that the correlations are
comparable.

The way the correlation function data is stored has also changed; all
data is parsed and manipulated using a new qmc::quantity data type that
automatically propagates error at the cost of some computational
efficiency, and the correlation functions are now array-indexed using a
new enumeration type. This allows us to cut down on the lines of code
needed at the cost of some slight readability overhead, though
hopefully not much.