Add flag to conditionally transform correl. fns
For debugging purposes, we used to transform n -> 2-n, etc. to be able
to compare mu > 0 and mu < 0 correlation data. This was primarily done
for bug-checking of the nup-nud and ndown-nud data, which was erroneous.
As such, we've now added flags to determine whether to do this
transformation at the moment the output files are written, allowing us
to switch between debugging data and production data at any time.