Add parsing of average z-spin and correl. fn
We've extended the size of the maps and structs to include objects for
the unaveraged/averaged z-spins and correlation functions. This means
the program also now outputs a _cs.dat file that contains Cs(r) data,
analogous to the Cm(r) data we printed before. For details on what this
correlation function looks like, see 1606.04089, specifically Fig. 2 as
well as Eq. 3 (and the paragraph before it).