Modularize a bit
This explicitly renames all the files relating to bandmott-specific
output handling as to not imply a preferred format for data parsing.
This doesn't really mean much, as the only currently support input *is*
the bandmott program's output files, but it sets the stage for the
future.

We also split the files into separate units... still getting a feel for
how to idiomatically handle this in Julia, though.